helm: Fix ServiceAccount for Scylla Manager chart
The deployment/statefulset relied on the "default" Service Account, but
didn't bind the right ClusterRole when another Service Account is used
as defined by the template helper `scylla-manager.serviceAccountName`.

This fix makes sure the ClusterRoleBinding is bound to the right Service
Account and makes sure to assign the proper Service Account to the
Deployment and Statefulset.

I've also updated the naming of the ClusterRole and ClusterRoleBinding
to be aligned with the same approach as for the Scylla chart